@@ -1,16 +1,16 @@
 name: fwfh_url_launcher
-version: 0.6.1+3
+version: 0.9.0+1
 description: WidgetFactory extension to launch A tag via url_launcher plugin.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=2.0.0"
-  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.2.0"
+  sdk: ">=2.13.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_widget_from_html_core: ">=0.6.1 <0.9.0"
+  flutter_widget_from_html_core: ">=0.8.0 <0.11.0"
   url_launcher: ^6.0.0
 
 dependency_overrides:
