@@ -190,7 +190,13 @@ class HtmlWidgetState extends State<HtmlWidget> {
             return snapshot.data!;
           } else if (snapshot.hasError) {
             return _sliverToBoxAdapterIfNeeded(
-              _wf.onErrorBuilder(context, _rootMeta, snapshot.error) ?? widget0,
+              _wf.onErrorBuilder(
+                    context,
+                    _rootMeta,
+                    snapshot.error,
+                    snapshot.stackTrace,
+                  ) ??
+                  widget0,
             );
           } else {
             return _sliverToBoxAdapterIfNeeded(
@@ -214,6 +220,9 @@ class HtmlWidgetState extends State<HtmlWidget> {
 
   Future<Widget> _buildAsync() async {
     final domNodes = await compute(_parseHtml, widget.html);
+    if (!mounted) {
+      return widget0;
+    }
 
     Timeline.startSync('Build $widget (async)');
     final built = _buildBody(this, domNodes);
@@ -225,8 +234,14 @@ class HtmlWidgetState extends State<HtmlWidget> {
   Widget _buildSync() {
     Timeline.startSync('Build $widget (sync)');
 
-    final domNodes = _parseHtml(widget.html);
-    final built = _buildBody(this, domNodes);
+    Widget built;
+    try {
+      final domNodes = _parseHtml(widget.html);
+      built = _buildBody(this, domNodes);
+    } catch (error, stackTrace) {
+      built =
+          _wf.onErrorBuilder(context, _rootMeta, error, stackTrace) ?? widget0;
+    }
 
     Timeline.finishSync();
 
