@@ -3,6 +3,8 @@ import 'dart:math';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
+import 'css_sizing.dart';
+
 /// A TABLE widget.
 class HtmlTable extends MultiChildRenderObjectWidget {
   /// The table border sides.
@@ -18,21 +20,22 @@ class HtmlTable extends MultiChildRenderObjectWidget {
   /// Default: `0.0`.
   final double borderSpacing;
 
-  /// The companion data for table.
-  final HtmlTableCompanion companion;
-
   /// Determines the order to lay children out horizontally.
   ///
   /// Default: [TextDirection.ltr].
   final TextDirection textDirection;
 
+  // TODO: remove lint ignore when our minimum Flutter version >= 3.10
+  // https://github.com/flutter/flutter/pull/119195
+  // https://github.com/flutter/flutter/commit/6a5405925dffb5b4121e1fba898d3d2068dac77c
+
   /// Creates a TABLE widget.
+  // ignore: prefer_const_constructors_in_immutables
   HtmlTable({
     this.border,
     this.borderCollapse = false,
     this.borderSpacing = 0.0,
     required List<Widget> children,
-    required this.companion,
     this.textDirection = TextDirection.ltr,
     Key? key,
   }) : super(children: children, key: key);
@@ -41,7 +44,6 @@ class HtmlTable extends MultiChildRenderObjectWidget {
   RenderObject createRenderObject(BuildContext _) => _TableRenderObject(
         border,
         borderSpacing,
-        companion,
         textDirection,
         borderCollapse: borderCollapse,
       );
@@ -72,17 +74,29 @@ class HtmlTable extends MultiChildRenderObjectWidget {
   @override
   void updateRenderObject(BuildContext _, RenderObject renderObject) {
     (renderObject as _TableRenderObject)
-      ..border = border
-      ..borderCollapse = borderCollapse
-      ..borderSpacing = borderSpacing
-      ..companion = companion
-      ..textDirection = textDirection;
+      ..setBorder(border)
+      ..setBorderCollapse(borderCollapse)
+      ..setBorderSpacing(borderSpacing)
+      ..setTextDirection(textDirection);
   }
 }
 
-/// Companion data for table.
-class HtmlTableCompanion {
-  final _baselines = <int, List<_ValignBaselineRenderObject>>{};
+/// A table caption widget.
+class HtmlTableCaption extends HtmlTableCell {
+  /// Creates a table caption widget.
+  const HtmlTableCaption({
+    required Widget child,
+    required int columnSpan,
+    required int rowIndex,
+    Key? key,
+  }) : super._(
+          columnStart: 0,
+          columnSpan: columnSpan,
+          isCaption: true,
+          key: key,
+          rowStart: rowIndex,
+          child: child,
+        );
 }
 
 /// A TD (table cell) widget.
@@ -102,19 +116,38 @@ class HtmlTableCell extends ParentDataWidget<_TableCellData> {
   /// The row index this cell should start.
   final int rowStart;
 
+  /// The cell width.
+  final CssSizingValue? width;
+
+  final bool _isCaption;
+
   /// Creates a TD (table cell) widget.
-  const HtmlTableCell({
+  const factory HtmlTableCell({
+    Border? border,
+    required Widget child,
+    int columnSpan,
+    required int columnStart,
+    Key? key,
+    int rowSpan,
+    required int rowStart,
+    CssSizingValue? width,
+  }) = HtmlTableCell._;
+
+  const HtmlTableCell._({
     this.border,
     required Widget child,
     this.columnSpan = 1,
     required this.columnStart,
+    bool isCaption = false,
     Key? key,
     this.rowSpan = 1,
     required this.rowStart,
+    this.width,
   })  : assert(columnSpan >= 1),
         assert(columnStart >= 0),
         assert(rowSpan >= 1),
         assert(rowStart >= 0),
+        _isCaption = isCaption,
         super(child: child, key: key);
 
   @override
@@ -137,6 +170,11 @@ class HtmlTableCell extends ParentDataWidget<_TableCellData> {
       needsLayout = true;
     }
 
+    if (data.isCaption != _isCaption) {
+      data.isCaption = _isCaption;
+      needsLayout = true;
+    }
+
     if (data.rowStart != rowStart) {
       data.rowStart = rowStart;
       needsLayout = true;
@@ -144,6 +182,11 @@ class HtmlTableCell extends ParentDataWidget<_TableCellData> {
 
     if (data.rowSpan != rowSpan) {
       data.rowSpan = rowSpan;
+      needsLayout = true;
+    }
+
+    if (data.width != width) {
+      data.width = width;
       needsLayout = true;
     }
 
@@ -169,49 +212,49 @@ class HtmlTableCell extends ParentDataWidget<_TableCellData> {
   Type get debugTypicalAncestorWidgetClass => HtmlTable;
 }
 
-/// A `valign=baseline` widget.
-class HtmlTableValignBaseline extends SingleChildRenderObjectWidget {
-  /// Creates a `valign=baseline` widget.
-  const HtmlTableValignBaseline({Widget? child, Key? key})
-      : super(child: child, key: key);
-
-  @override
-  RenderObject createRenderObject(BuildContext context) {
-    final table = context.findAncestorWidgetOfExactType<HtmlTable>()!;
-    final cell = context.findAncestorWidgetOfExactType<HtmlTableCell>()!;
-    return _ValignBaselineRenderObject(table.companion, cell.rowStart);
-  }
-
-  @override
-  void updateRenderObject(BuildContext context, RenderObject renderObject) {
-    final table = context.findAncestorWidgetOfExactType<HtmlTable>()!;
-    final cell = context.findAncestorWidgetOfExactType<HtmlTableCell>()!;
-    (renderObject as _ValignBaselineRenderObject)
-      ..companion = table.companion
-      ..row = cell.rowStart;
-  }
-}
-
 extension _IterableDouble on Iterable<double> {
   double get sum => isEmpty ? 0.0 : reduce(_sum);
 
   static double _sum(double value, double element) => value + element;
 }
 
+extension _IterableDoubleList on List<double> {
+  void setMaxColumnWidths(
+    _TableRenderObject tro,
+    _TableCellData data,
+    double calculatedWidth,
+  ) {
+    final columnWidth =
+        (calculatedWidth - tro._calculateColumnGaps(data)) / data.columnSpan;
+    for (var c = 0; c < data.columnSpan; c++) {
+      final column = data.columnStart + c;
+
+      // make sure each column has enough width for the child
+      // oversimplified: the colspan splits its columns evenly
+      this[column] = max(this[column], columnWidth);
+    }
+  }
+
+  double sumRange(_TableCellData data) =>
+      getRange(data.columnStart, data.columnStart + data.columnSpan).sum;
+}
+
 class _TableCellData extends ContainerBoxParentData<RenderBox> {
   Border? border;
   int columnSpan = 1;
-  int columnStart = -1;
+  int columnStart = 0;
+  bool isCaption = false;
   int rowSpan = 1;
-  int rowStart = -1;
+  int rowStart = 0;
+  CssSizingValue? width;
 
   double calculateHeight(_TableRenderObject tro, List<double> heights) {
-    final gaps = (rowSpan - 1) * tro.rowGap;
+    final gaps = tro._calculateRowGaps(this);
     return heights.getRange(rowStart, rowStart + rowSpan).sum + gaps;
   }
 
   double calculateWidth(_TableRenderObject tro, List<double> widths) {
-    final gaps = (columnSpan - 1) * tro.columnGap;
+    final gaps = tro._calculateColumnGaps(this);
     return widths.getRange(columnStart, columnStart + columnSpan).sum + gaps;
   }
 
@@ -222,9 +265,311 @@ class _TableCellData extends ContainerBoxParentData<RenderBox> {
   }
 
   double calculateY(_TableRenderObject tro, List<double> heights) {
-    final padding = tro._border?.top.width ?? 0.0;
+    final padding = isCaption ? .0 : (tro._border?.top.width ?? 0.0);
     final gaps = (rowStart + 1) * tro.rowGap;
     return padding + heights.getRange(0, rowStart).sum + gaps;
+  }
+}
+
+class _TableRenderLayout {
+  final Rect cellRect;
+  final Size totalSize;
+
+  const _TableRenderLayout(this.cellRect, this.totalSize);
+
+  static const _TableRenderLayout zero =
+      _TableRenderLayout(Rect.zero, Size.zero);
+}
+
+class _TableRenderLayouter {
+  final BoxConstraints constraints;
+  final Size Function(RenderBox renderBox, BoxConstraints constraints) layouter;
+  final _TableRenderObject tro;
+
+  final List<_TableCellData> cells = [];
+  final List<RenderBox> children = [];
+  final List<Size> drySizes = [];
+
+  double captionHeight = .0;
+  int columnCount = 0;
+  int rowCount = 0;
+
+  _TableRenderLayouter(this.tro, this.constraints)
+      : layouter = performLayoutLayout;
+
+  _TableRenderLayouter.dry(this.tro, this.constraints)
+      : layouter = performLayoutGetDryLayout;
+
+  _TableRenderLayout layout(RenderBox firstChild) {
+    step1GuessDrySizes(firstChild);
+
+    final columnGapsSum = (columnCount + 1) * tro.columnGap;
+    final gapsAndPaddings = tro.paddingLeft + columnGapsSum + tro.paddingRight;
+    final availableWidth = constraints.maxWidth - gapsAndPaddings;
+    final dryColumnWidths = step2DryColumnWidths();
+
+    final columnWidths = step3ColumnWiths(availableWidth, dryColumnWidths);
+
+    final childSizes = List.filled(children.length, Size.zero);
+    final rowHeights = List.filled(rowCount, .0);
+    step4ChildSizesAndRowHeights(columnWidths, childSizes, rowHeights);
+
+    return step5CalculateChildrenOffsetMaybeRelayout(
+      childSizes,
+      columnWidths,
+      rowHeights,
+    );
+  }
+
+  void step1GuessDrySizes(RenderBox firstChild) {
+    RenderBox? child = firstChild;
+    while (child != null) {
+      final data = child.parentData! as _TableCellData;
+      children.add(child);
+      cells.add(data);
+
+      Size drySize;
+      final width = data.width?.clamp(0, constraints.maxWidth);
+      if (width != null) {
+        drySize = Size(width, .0);
+      } else {
+        drySize =
+            Size(constraints.hasBoundedWidth ? constraints.maxWidth : 100.0, 0);
+        try {
+          drySize = performLayoutGetDryLayout(child, const BoxConstraints());
+        } catch (dryLayoutError, stackTrace) {
+          debugPrint('Ignored _performLayoutDry error: '
+              '$dryLayoutError\n$stackTrace');
+        }
+      }
+      drySizes.add(drySize);
+
+      columnCount = max(columnCount, data.columnStart + data.columnSpan);
+      rowCount = max(rowCount, data.rowStart + data.rowSpan);
+
+      child = data.nextSibling;
+    }
+  }
+
+  List<double> step2DryColumnWidths() {
+    final dryColumnWidths = List.filled(columnCount, .0);
+    for (var i = 0; i < children.length; i++) {
+      final data = children[i].parentData! as _TableCellData;
+      final drySize = drySizes[i];
+      dryColumnWidths.setMaxColumnWidths(tro, data, drySize.width);
+    }
+
+    return dryColumnWidths;
+  }
+
+  List<double> step3ColumnWiths(
+    double availableWidth,
+    List<double> dryColumnWidths,
+  ) {
+    // being naive: take dry widths as render widths
+    var columnWidths = [...dryColumnWidths];
+    if (availableWidth.isInfinite || dryColumnWidths.sum <= availableWidth) {
+      return columnWidths;
+    }
+
+    final minColumnWidths = List.filled(columnCount, 0.0);
+    var shouldLoop = true;
+    var loopCount = 0;
+    while (shouldLoop) {
+      shouldLoop = false;
+
+      columnWidths =
+          redistributeValues(columnWidths, minColumnWidths, availableWidth);
+
+      for (var i = 0; i < children.length; i++) {
+        final child = children[i];
+        final data = cells[i];
+        final drySize = drySizes[i];
+
+        final dataGaps = tro._calculateColumnGaps(data);
+        final inCalculationWidth = columnWidths.sumRange(data) + dataGaps;
+        if (drySize.width > inCalculationWidth) {
+          double? minWidth;
+          try {
+            // this call is expensive, we try to avoid it as much as possible
+            // width being smaller than dry size means the table is too crowded
+            // calculating min to avoid breaking line in the middle of a word
+            minWidth = child.getMinIntrinsicWidth(double.infinity);
+          } catch (minWidthError, stackTrace) {
+            minWidth = drySize.width;
+            debugPrint('Ignored getMinIntrinsicWidth error: '
+                '$minWidthError\n$stackTrace');
+          }
+
+          minColumnWidths.setMaxColumnWidths(tro, data, minWidth);
+
+          // side effect
+          // keep track of min width in the array to avoid
+          // processing the same child more than once
+          drySizes[i] = Size(minWidth, 0);
+
+          // the loop should run at least one more time with new min-width
+          shouldLoop = true;
+        }
+      }
+
+      loopCount++;
+      if (loopCount > columnCount) {
+        // using column count to stop early, not a typo
+        // we don't want to waste too much time in this loop
+        // in case the table is extra long with many many rows
+        debugPrint('Stopped to avoid infinite loops latest=$columnWidths');
+        break;
+      }
+    }
+
+    return columnWidths;
+  }
+
+  void step4ChildSizesAndRowHeights(
+    List<double> columnWidths,
+    List<Size> childSizes,
+    List<double> rowHeights,
+  ) {
+    for (var i = 0; i < children.length; i++) {
+      final child = children[i];
+      final data = cells[i];
+      final drySize = drySizes[i];
+
+      final childWidth = data.calculateWidth(tro, columnWidths);
+      final canUseDrySize = childWidth == drySize.width &&
+          drySize.height > 0 &&
+          identical(layouter, performLayoutGetDryLayout);
+      Size childSize;
+      if (canUseDrySize) {
+        childSize = drySize;
+      } else {
+        // side effect
+        childSize = layouter(child, BoxConstraints.tightFor(width: childWidth));
+      }
+      childSizes[i] = childSize;
+
+      // distribute cell height across spanned rows
+      final rowHeight =
+          (childSize.height - tro._calculateRowGaps(data)) / data.rowSpan;
+      for (var r = 0; r < data.rowSpan; r++) {
+        final row = data.rowStart + r;
+        rowHeights[row] = max(rowHeights[row], rowHeight);
+      }
+    }
+  }
+
+  _TableRenderLayout step5CalculateChildrenOffsetMaybeRelayout(
+    List<Size> childSizes,
+    List<double> columnWidths,
+    List<double> rowHeights,
+  ) {
+    final columnGapsSum = (columnCount + 1) * tro.columnGap;
+    final rowGapsSum = (rowCount + 1) * tro.rowGap;
+    final calculatedHeight =
+        tro.paddingTop + rowHeights.sum + rowGapsSum + tro.paddingBottom;
+    final constraintedHeight = constraints.constrainHeight(calculatedHeight);
+    final deltaHeight =
+        max(0, (constraintedHeight - calculatedHeight) / rowCount);
+    final calculatedWidth =
+        tro.paddingLeft + columnWidths.sum + columnGapsSum + tro.paddingRight;
+    for (var i = 0; i < children.length; i++) {
+      final child = children[i];
+      final data = cells[i];
+      var childSize = childSizes[i];
+
+      var childHeight = data.calculateHeight(tro, rowHeights) + deltaHeight;
+      var childWidth = data.calculateWidth(tro, columnWidths);
+      if (childSize.height != childHeight) {
+        final cc2 = BoxConstraints.tight(Size(childWidth, childHeight));
+        // side effect
+        childSize = layouter(child, cc2);
+        childHeight = childSize.height;
+        childWidth = childSize.width;
+      }
+
+      final calculatedY = data.calculateY(tro, rowHeights);
+      if (child.hasSize) {
+        final calculatedX = data.calculateX(tro, columnWidths);
+
+        double x;
+        switch (tro._textDirection) {
+          case TextDirection.ltr:
+            x = calculatedX;
+            break;
+          case TextDirection.rtl:
+            x = calculatedWidth - childWidth - calculatedX;
+            break;
+        }
+
+        data.offset = Offset(x, calculatedY);
+      }
+
+      if (data.isCaption) {
+        captionHeight = max(captionHeight, calculatedY + childHeight);
+      }
+    }
+
+    return _TableRenderLayout(
+      Rect.fromLTWH(
+        0,
+        captionHeight,
+        calculatedWidth,
+        calculatedHeight - captionHeight,
+      ),
+      Size(calculatedWidth, calculatedHeight),
+    );
+  }
+
+  static Size performLayoutGetDryLayout(
+    RenderBox renderBox,
+    BoxConstraints constraints,
+  ) =>
+      renderBox.getDryLayout(constraints);
+
+  static Size performLayoutLayout(
+    RenderBox renderBox,
+    BoxConstraints constraints,
+  ) {
+    renderBox.layout(constraints, parentUsesSize: true);
+    return renderBox.size;
+  }
+
+  static List<double> redistributeValues(
+    List<double> values,
+    List<double> calculatedMinValues,
+    double available,
+  ) {
+    final effectiveMinValues = List.filled(values.length, .0);
+    for (var i = 0; i < values.length; i++) {
+      if (calculatedMinValues[i] > 0 && calculatedMinValues[i] >= values[i]) {
+        // min value is smaller than in-calculation width
+        // let's keep the current value as long as possible
+        // we want the column to grow to its dry size naturally
+        effectiveMinValues[i] = calculatedMinValues[i];
+      }
+    }
+
+    final remaining = max(.0, available - effectiveMinValues.sum);
+    var valuesSum = .0;
+    for (var i = 0; i < values.length; i++) {
+      if (effectiveMinValues[i] == 0) {
+        valuesSum += values[i];
+      }
+    }
+
+    final result = effectiveMinValues.toList(growable: false);
+    if (valuesSum > .0) {
+      for (var i = 0; i < values.length; i++) {
+        if (result[i] == 0) {
+          // calculate widths using weighted distribution
+          // e.g. if a column has huge dry width, it will have bigger width
+          result[i] = values[i] / valuesSum * remaining;
+        }
+      }
+    }
+
+    return result;
   }
 }
 
@@ -235,64 +580,43 @@ class _TableRenderObject extends RenderBox
   _TableRenderObject(
     this._border,
     this._borderSpacing,
-    this._companion,
     this._textDirection, {
     required bool borderCollapse,
   }) : _borderCollapse = borderCollapse;
 
   Border? _border;
-  // ignore: avoid_setters_without_getters
-  set border(Border? v) {
-    if (v == _border) {
-      return;
+  void setBorder(Border? v) {
+    if (v != _border) {
+      _border = v;
+      markNeedsLayout();
     }
-
-    _border = v;
-    markNeedsLayout();
   }
 
   bool _borderCollapse;
-  // ignore: avoid_setters_without_getters
-  set borderCollapse(bool v) {
-    if (v == _borderCollapse) {
-      return;
+  // ignore: avoid_positional_boolean_parameters
+  void setBorderCollapse(bool v) {
+    if (v != _borderCollapse) {
+      _borderCollapse = v;
+      markNeedsLayout();
     }
-
-    _borderCollapse = v;
-    markNeedsLayout();
   }
 
   double _borderSpacing;
-  // ignore: avoid_setters_without_getters
-  set borderSpacing(double v) {
-    if (v == _borderSpacing) {
-      return;
+  void setBorderSpacing(double v) {
+    if (v != _borderSpacing) {
+      _borderSpacing = v;
+      markNeedsLayout();
     }
-
-    _borderSpacing = v;
-    markNeedsLayout();
   }
 
-  HtmlTableCompanion _companion;
-  // ignore: avoid_setters_without_getters
-  set companion(HtmlTableCompanion v) {
-    if (v == _companion) {
-      return;
-    }
-
-    _companion = v;
-    markNeedsLayout();
-  }
+  var _layout = _TableRenderLayout.zero;
 
   TextDirection _textDirection;
-  // ignore: avoid_setters_without_getters
-  set textDirection(TextDirection v) {
-    if (v == _textDirection) {
-      return;
+  void setTextDirection(TextDirection v) {
+    if (v != _textDirection) {
+      _textDirection = v;
+      markNeedsLayout();
     }
-
-    _textDirection = v;
-    markNeedsLayout();
   }
 
   double get columnGap => _border != null && _borderCollapse
@@ -310,9 +634,6 @@ class _TableRenderObject extends RenderBox
   double get rowGap => _border != null && _borderCollapse
       ? (_border!.top.width * -1.0)
       : _borderSpacing;
-
-  double? _calculatedHeight;
-  double? _calculatedWidth;
 
   @override
   double? computeDistanceToActualBaseline(TextBaseline baseline) {
@@ -345,8 +666,10 @@ class _TableRenderObject extends RenderBox
   }
 
   @override
-  Size computeDryLayout(BoxConstraints constraints) =>
-      _performLayout(this, firstChild!, constraints, _performLayoutDry);
+  Size computeDryLayout(BoxConstraints constraints) {
+    final layouter = _TableRenderLayouter.dry(this, constraints);
+    return layouter.layout(firstChild!).totalSize;
+  }
 
   @override
   bool hitTestChildren(BoxHitTestResult result, {required Offset position}) =>
@@ -354,36 +677,20 @@ class _TableRenderObject extends RenderBox
 
   @override
   void paint(PaintingContext context, Offset offset) {
-    _companion._baselines.clear();
-
-    assert(_calculatedHeight != null);
-    assert(_calculatedWidth != null);
-    _border?.paint(
-      context.canvas,
-      Rect.fromLTWH(
-        offset.dx,
-        offset.dy,
-        _calculatedWidth ?? 0.0,
-        _calculatedHeight ?? 0.0,
-      ),
-    );
+    final cellRect = _layout.cellRect;
+    if (!cellRect.isEmpty) {
+      _border?.paint(context.canvas, cellRect.shift(offset));
+    }
 
     var child = firstChild;
     while (child != null) {
       final data = child.parentData! as _TableCellData;
-      final childOffset = data.offset + offset;
-      final childSize = child.size;
-      context.paintChild(child, childOffset);
+      final o = data.offset + offset;
+      final s = child.size;
+      context.paintChild(child, o);
 
-      data.border?.paint(
-        context.canvas,
-        Rect.fromLTWH(
-          childOffset.dx,
-          childOffset.dy,
-          childSize.width,
-          childSize.height,
-        ),
-      );
+      final borderRect = Rect.fromLTWH(o.dx, o.dy, s.width, s.height);
+      data.border?.paint(context.canvas, borderRect);
 
       child = data.nextSibling;
     }
@@ -391,12 +698,9 @@ class _TableRenderObject extends RenderBox
 
   @override
   void performLayout() {
-    final calculatedSize =
-        _performLayout(this, firstChild!, constraints, _performLayoutLayouter);
-    _calculatedHeight = calculatedSize.height;
-    _calculatedWidth = calculatedSize.width;
-
-    size = constraints.constrain(calculatedSize);
+    final layouter = _TableRenderLayouter(this, constraints);
+    _layout = layouter.layout(firstChild!);
+    size = constraints.constrain(_layout.totalSize);
   }
 
   @override
@@ -406,251 +710,11 @@ class _TableRenderObject extends RenderBox
     }
   }
 
-  static Size _performLayout(
-    final _TableRenderObject tro,
-    final RenderBox firstChild,
-    final BoxConstraints constraints,
-    final Size Function(RenderBox renderBox, BoxConstraints constraints)
-        layouter,
-  ) {
-    final children = <RenderBox>[];
-    final cells = <_TableCellData>[];
-
-    RenderBox? child = firstChild;
-    var columnCount = 0;
-    var rowCount = 0;
-    while (child != null) {
-      final data = child.parentData! as _TableCellData;
-      children.add(child);
-      cells.add(data);
-
-      columnCount = max(columnCount, data.columnStart + data.columnSpan);
-      rowCount = max(rowCount, data.rowStart + data.rowSpan);
-
-      child = data.nextSibling;
-    }
-
-    final columnGaps = (columnCount + 1) * tro.columnGap;
-    final rowGaps = (rowCount + 1) * tro.rowGap;
-    final width0 = (constraints.maxWidth -
-            tro.paddingLeft -
-            tro.paddingRight -
-            columnGaps) /
-        columnCount;
-    final childSizes = List.filled(children.length, Size.zero);
-    final columnWidths = List.filled(columnCount, 0.0);
-    final rowHeights = List.filled(rowCount, 0.0);
-    for (var i = 0; i < children.length; i++) {
-      final child = children[i];
-      final data = cells[i];
-
-      // assume even distribution of column widths if width is finite
-      final childColumnGaps = (data.columnSpan - 1) * tro.columnGap;
-      final childWidth =
-          width0.isFinite ? width0 * data.columnSpan + childColumnGaps : null;
-      final cc = BoxConstraints(
-        maxWidth: childWidth ?? double.infinity,
-        minWidth: childWidth ?? 0.0,
-      );
-      final childSize = childSizes[i] = layouter(child, cc);
-
-      // distribute cell width across spanned columns
-      final columnWidth = (childSize.width - childColumnGaps) / data.columnSpan;
-      for (var c = 0; c < data.columnSpan; c++) {
-        final column = data.columnStart + c;
-        columnWidths[column] = max(columnWidths[column], columnWidth);
-      }
-
-      // distribute cell height across spanned rows
-      final childRowGaps = (data.rowSpan - 1) * tro.rowGap;
-      final rowHeight = (childSize.height - childRowGaps) / data.rowSpan;
-      for (var r = 0; r < data.rowSpan; r++) {
-        final row = data.rowStart + r;
-        rowHeights[row] = max(rowHeights[row], rowHeight);
-      }
-    }
-
-    // we now know all the widths and heights, let's position cells
-    // sometime we have to relayout child, e.g. stretch its height for rowspan
-    final calculatedHeight =
-        tro.paddingTop + rowHeights.sum + rowGaps + tro.paddingBottom;
-    final constraintedHeight = constraints.constrainHeight(calculatedHeight);
-    final deltaHeight =
-        max(0, (constraintedHeight - calculatedHeight) / rowCount);
-    final calculatedWidth =
-        tro.paddingLeft + columnWidths.sum + columnGaps + tro.paddingRight;
-    final constraintedWidth = constraints.constrainWidth(calculatedWidth);
-    final deltaWidth = (constraintedWidth - calculatedWidth) / columnCount;
-    for (var i = 0; i < children.length; i++) {
-      final child = children[i];
-      final data = cells[i];
-      final childSize = childSizes[i];
-
-      final childHeight = data.calculateHeight(tro, rowHeights) + deltaHeight;
-      final childWidth = data.calculateWidth(tro, columnWidths) + deltaWidth;
-      if (childSize.height != childHeight || childSize.width != childWidth) {
-        final cc2 = BoxConstraints.tight(Size(childWidth, childHeight));
-        layouter(child, cc2);
-      }
-
-      if (child.hasSize) {
-        final calculatedX = data.calculateX(tro, columnWidths);
-
-        // TODO: use `late final` when https://github.com/dart-lang/coverage/issues/341 is fixed
-        late double x;
-        switch (tro._textDirection) {
-          case TextDirection.ltr:
-            x = calculatedX;
-            break;
-          case TextDirection.rtl:
-            x = calculatedWidth - childWidth - calculatedX;
-            break;
-        }
-
-        final y = data.calculateY(tro, rowHeights);
-        data.offset = Offset(x, y);
-      }
-    }
-
-    return Size(calculatedWidth, calculatedHeight);
+  double _calculateColumnGaps(_TableCellData data) {
+    return (data.columnSpan - 1) * columnGap;
   }
 
-  static Size _performLayoutDry(
-    RenderBox renderBox,
-    BoxConstraints constraints,
-  ) =>
-      renderBox.getDryLayout(constraints);
-
-  static Size _performLayoutLayouter(
-    RenderBox renderBox,
-    BoxConstraints constraints,
-  ) {
-    renderBox.layout(constraints, parentUsesSize: true);
-    return renderBox.size;
-  }
-}
-
-class _ValignBaselineRenderObject extends RenderProxyBox {
-  _ValignBaselineRenderObject(this._companion, this._row);
-
-  HtmlTableCompanion _companion;
-  // ignore: avoid_setters_without_getters
-  set companion(HtmlTableCompanion v) {
-    if (v == _companion) {
-      return;
-    }
-    _companion = v;
-    markNeedsLayout();
-  }
-
-  int _row;
-  // ignore: avoid_setters_without_getters
-  set row(int v) {
-    if (v == _row) {
-      return;
-    }
-    _row = v;
-    markNeedsLayout();
-  }
-
-  double? _baselineWithOffset;
-  var _paddingTop = 0.0;
-
-  @override
-  Size computeDryLayout(BoxConstraints constraints) =>
-      _performLayout(child, _paddingTop, constraints, _performLayoutDry);
-
-  @override
-  void paint(PaintingContext context, Offset offset) {
-    final effectiveOffset = offset.translate(0, _paddingTop);
-
-    final child = this.child;
-    if (child == null) {
-      return;
-    }
-
-    final baselineWithOffset = _baselineWithOffset = effectiveOffset.dy +
-        (child.getDistanceToBaseline(TextBaseline.alphabetic) ?? 0.0);
-
-    final siblings = _companion._baselines;
-    if (siblings.containsKey(_row)) {
-      final rowBaseline = siblings[_row]!
-          .map((e) => e._baselineWithOffset!)
-          .reduce((v, e) => max(v, e));
-      siblings[_row]!.add(this);
-
-      if (rowBaseline > baselineWithOffset) {
-        final offsetY = rowBaseline - baselineWithOffset;
-        if (size.height - child.size.height >= offsetY) {
-          // paint child with additional offset
-          context.paintChild(child, effectiveOffset.translate(0, offsetY));
-          return;
-        } else {
-          // skip painting this frame, wait for the correct padding
-          _paddingTop += offsetY;
-          _baselineWithOffset = rowBaseline;
-          WidgetsBinding.instance
-              ?.addPostFrameCallback((_) => markNeedsLayout());
-          return;
-        }
-      } else if (rowBaseline < baselineWithOffset) {
-        for (final sibling in siblings[_row]!) {
-          if (sibling == this) {
-            continue;
-          }
-
-          final offsetY = baselineWithOffset - sibling._baselineWithOffset!;
-          if (offsetY != 0.0) {
-            sibling._paddingTop += offsetY;
-            sibling._baselineWithOffset = baselineWithOffset;
-            WidgetsBinding.instance
-                ?.addPostFrameCallback((_) => sibling.markNeedsLayout());
-          }
-        }
-      }
-    } else {
-      siblings[_row] = [this];
-    }
-
-    context.paintChild(child, effectiveOffset);
-  }
-
-  @override
-  void performLayout() {
-    size = _performLayout(
-      child,
-      _paddingTop,
-      constraints,
-      _performLayoutLayouter,
-    );
-  }
-
-  @override
-  String toStringShort() => '_ValignBaselineRenderObject(row: $_row)';
-
-  static Size _performLayout(
-    final RenderBox? child,
-    final double paddingTop,
-    final BoxConstraints constraints,
-    final Size? Function(RenderBox? renderBox, BoxConstraints constraints)
-        layouter,
-  ) {
-    final cc = constraints.loosen().deflate(EdgeInsets.only(top: paddingTop));
-    final childSize = layouter(child, cc) ?? Size.zero;
-    return constraints.constrain(childSize + Offset(0, paddingTop));
-  }
-
-  static Size? _performLayoutDry(
-    RenderBox? renderBox,
-    BoxConstraints constraints,
-  ) =>
-      renderBox?.getDryLayout(constraints);
-
-  static Size? _performLayoutLayouter(
-    RenderBox? renderBox,
-    BoxConstraints constraints,
-  ) {
-    renderBox?.layout(constraints, parentUsesSize: true);
-    return renderBox?.size;
+  double _calculateRowGaps(_TableCellData data) {
+    return (data.rowSpan - 1) * rowGap;
   }
 }
